cleanup of resource managenet, added few doccomments
