+++ src/DotVVM.Framework/ResourceManagement/CachingResourceRepository.cs
@@ -6,6 +6,9 @@ using System.Threading.Tasks;
 
 namespace DotVVM.Framework.ResourceManagement
 {
+    /// <summary>
+    /// Base <see cref="IDotvvmResourceRepository"/>. Ensures that FindResource is called only once for one name and rembers the result for next calls.
+    /// </summary>
     public abstract class CachingResourceRepository: IDotvvmResourceRepository
     {
         protected abstract IResource FindResource(string name);
