+++ src/DotVVM.Framework/ResourceManagement/EmbeddedResourceLocation.cs
@@ -1,0 +1,34 @@
+﻿using System.IO;
+using DotVVM.Framework.Hosting;
+using System.Reflection;
+using Newtonsoft.Json;
+
+namespace DotVVM.Framework.ResourceManagement
+{
+    /// <summary>
+    /// Represents resource located in assembly embedded resources.
+    /// </summary>
+    public class EmbeddedResourceLocation : LocalResourceLocation, IDebugFileLocalLocation
+    {
+        [JsonConverter(typeof(ReflectionAssemblyJsonConverter))]
+        public Assembly Assembly { get; }
+        public string Name { get; }
+        /// <summary>
+        /// File where the resource is located for debug purposes
+        /// </summary>
+        public string DebugFilePath { get; set; }
+        public EmbeddedResourceLocation(Assembly assembly, string name, string debugFileName = null)
+        {
+            this.Name = name;
+            this.Assembly = assembly;
+            this.DebugFilePath = debugFileName;
+        }
+
+        public override Stream LoadResource(IDotvvmRequestContext context) =>
+            context.Configuration.Debug && DebugFilePath != null && File.Exists(DebugFilePath) ?
+            File.OpenRead(Path.Combine(context.Configuration.ApplicationPhysicalPath, DebugFilePath)) :
+            Assembly.GetManifestResourceStream(Name);
+
+        public string GetFilePath(IDotvvmRequestContext context) => DebugFilePath;
+    }
+}
