+++ src/DotVVM.Framework/ResourceManagement/ILocalResourceLocation.cs
@@ -7,16 +7,11 @@ using System.Threading.Tasks;
 
 namespace DotVVM.Framework.ResourceManagement
 {
+    /// <summary>
+    /// Represents resource located on the server, so it can be loaded to stream easily.
+    /// </summary>
     public interface ILocalResourceLocation: IResourceLocation
     {
         Stream LoadResource(IDotvvmRequestContext context);
-    }
-
-    /// <summary>
-    /// Can get physical location of the file for debugging purposes. In that directory can be located asociated source maps and based on file will be the resource refreshed.
-    /// </summary>
-    public interface IDebugFileLocalLocation: ILocalResourceLocation
-    {
-        string GetFilePath(IDotvvmRequestContext context);
     }
 }
