+++ src/DotVVM.Framework/ResourceManagement/IResource.cs
@@ -7,6 +7,9 @@ using System.Threading.Tasks;
 
 namespace DotVVM.Framework.ResourceManagement
 {
+    /// <summary>
+    /// Represents a resources that can claimed from html page.
+    /// </summary>
     public interface IResource
     {
         void Render(IHtmlWriter writer, IDotvvmRequestContext context, string resourceName);
