+++ src/DotVVM.Framework/ResourceManagement/IResourceLocation.cs
@@ -6,6 +6,9 @@ using System.Threading.Tasks;
 
 namespace DotVVM.Framework.ResourceManagement
 {
+    /// <summary>
+    /// Represents a location where resource can be found.
+    /// </summary>
     public interface IResourceLocation
     {
         string GetUrl(IDotvvmRequestContext context, string name);
