+++ src/DotVVM.Framework/ResourceManagement/IResourceProcessor.cs
@@ -6,6 +6,9 @@ using System.Threading.Tasks;
 
 namespace DotVVM.Framework.ResourceManagement
 {
+    /// <summary>
+    /// Post-processes requested resource by a page.
+    /// </summary>
     public interface IResourceProcessor
     {
         IEnumerable<NamedResource> Process(IEnumerable<NamedResource> source);
